requirements.txt
streamlit
langchain
langchain-community
langchain-groq
langchain-text-splitters
python-dotenv
pypdf
faiss-cpu
sentence-transformers
torch